@@ -177,7 +177,7 @@ class _SplitViewState extends State<SplitView> {
                   : Center(
                       child: Text(
                         'Pick a channel to show the messages 💬',
-                        style: Theme.of(context).textTheme.headline5,
+                        style: Theme.of(context).textTheme.headlineSmall,
                       ),
                     ),
             ),
@@ -300,6 +300,7 @@ class _ChannelPageState extends State<ChannelPage> {
   @override
   void dispose() {
     focusNode.dispose();
+    messageInputController.dispose();
     super.dispose();
   }
 }
