@@ -22,6 +22,7 @@ class StreamImageAttachment extends StreamAttachmentWidget {
     this.imageThumbnailSize = const Size(400, 400),
     this.imageThumbnailResizeType = 'clip',
     this.imageThumbnailCropType = 'center',
+    this.attachmentActionsModalBuilder,
   });
 
   /// The [StreamMessageThemeData] to use for the image title
@@ -52,26 +53,40 @@ class StreamImageAttachment extends StreamAttachmentWidget {
   /// Defaults to [center]
   final String /*center|top|bottom|left|right*/ imageThumbnailCropType;
 
+  /// {@macro attachmentActionsBuilder}
+  final AttachmentActionsBuilder? attachmentActionsModalBuilder;
+
   @override
   Widget build(BuildContext context) {
     return source.when(
       local: () {
-        if (attachment.localUri == null || attachment.file?.bytes == null) {
-          return AttachmentError(constraints: constraints);
-        }
-        return _buildImageAttachment(
-          context,
-          Image.memory(
-            attachment.file!.bytes!,
-            height: constraints?.maxHeight,
-            width: constraints?.maxWidth,
-            fit: BoxFit.cover,
-            errorBuilder: (context, _, __) => Image.asset(
-              'images/placeholder.png',
-              package: 'stream_chat_flutter',
+        if (attachment.file?.bytes != null) {
+          return _buildImageAttachment(
+            context,
+            Image.memory(
+              attachment.file!.bytes!,
+              height: constraints?.maxHeight,
+              width: constraints?.maxWidth,
+              fit: BoxFit.cover,
+              errorBuilder: _imageErrorBuilder,
             ),
-          ),
-        );
+          );
+        } else if (attachment.localUri != null) {
+          return _buildImageAttachment(
+            context,
+            Image.asset(
+              attachment.localUri!.path,
+              height: constraints?.maxHeight,
+              width: constraints?.maxWidth,
+              fit: BoxFit.cover,
+              errorBuilder: _imageErrorBuilder,
+            ),
+          );
+        } else {
+          return AttachmentError(
+            constraints: constraints,
+          );
+        }
       },
       network: () {
         var imageUrl =
@@ -116,6 +131,12 @@ class StreamImageAttachment extends StreamAttachmentWidget {
     );
   }
 
+  Widget _imageErrorBuilder(BuildContext _, Object __, StackTrace? ___) =>
+      Image.asset(
+        'images/placeholder.png',
+        package: 'stream_chat_flutter',
+      );
+
   Widget _buildImageAttachment(BuildContext context, Widget imageWidget) {
     return Container(
       constraints: constraints,
@@ -144,6 +165,8 @@ class StreamImageAttachment extends StreamAttachmentWidget {
                                     userName: message.user!.name,
                                     onShowMessage: onShowMessage,
                                     onReplyMessage: onReplyMessage,
+                                    attachmentActionsModalBuilder:
+                                        attachmentActionsModalBuilder,
                                   ),
                                 );
                               },
